# docs/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SLADashboardKPI": {
        "properties": {
          "net_payout": {
            "title": "Net Payout",
            "type": "number"
          },
          "total_outages": {
            "minimum": 0.0,
            "title": "Total Outages",
            "type": "integer"
          },
          "total_penalties": {
            "minimum": 0.0,
            "title": "Total Penalties",
            "type": "number"
          },
          "total_rewards": {
            "minimum": 0.0,
            "title": "Total Rewards",
            "type": "number"
          },
          "total_violations": {
            "minimum": 0.0,
            "title": "Total Violations",
            "type": "integer"
          }
        },
        "required": [
          "total_outages",
          "total_violations",
          "total_rewards",
          "total_penalties",
          "net_payout"
        ],
        "title": "SLADashboardKPI",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "ApexChainx Backend API",
    "title": "ApexChainx API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sla/analytics/dashboard": {
      "get": {
        "operationId": "get_sla_dashboard_kpis_api_v1_sla_analytics_dashboard_get",
        "parameters": [
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "description": "Alias for site_id",
            "in": "query",
            "name": "site",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Alias for site_id",
              "title": "Site"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SLADashboardKPI"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sla Dashboard Kpis",
        "tags": [
          "sla"
        ]
      }
    },
    "/api/v1/sla/analytics/dashboard/export": {
      "get": {
        "description": "Export dashboard KPI data in JSON or CSV format.",
        "operationId": "export_dashboard_kpis_api_v1_sla_analytics_dashboard_export_get",
        "parameters": [
          {
            "description": "Export format: json or csv",
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "description": "Export format: json or csv",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "description": "Alias for site_id",
            "in": "query",
            "name": "site",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Alias for site_id",
              "title": "Site"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Dashboard Kpis",
        "tags": [
          "sla"
        ]
      }
    },
    "/api/v1/sla/analytics/export": {
      "get": {
        "description": "Export comprehensive analytics summary (KPI + trends + optional aggregation).",
        "operationId": "export_analytics_summary_endpoint_api_v1_sla_analytics_export_get",
        "parameters": [
          {
            "description": "Export format: json or csv",
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "description": "Export format: json or csv",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "description": "Number of days for trends",
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "description": "Number of days for trends",
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "description": "Bucket interval: day, week, month",
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "default": "day",
              "description": "Bucket interval: day, week, month",
              "title": "Bucket",
              "type": "string"
            }
          },
          {
            "description": "IANA timezone name",
            "in": "query",
            "name": "tz",
            "required": false,
            "schema": {
              "default": "UTC",
              "description": "IANA timezone name",
              "title": "Tz",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "description": "Alias for site_id",
            "in": "query",
            "name": "site",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Alias for site_id",
              "title": "Site"
            }
          },
          {
            "description": "Include performance aggregation",
            "in": "query",
            "name": "include_aggregation",
            "required": false,
            "schema": {
              "default": true,
              "description": "Include performance aggregation",
              "title": "Include Aggregation",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Analytics Summary Endpoint",
        "tags": [
          "sla"
        ]
      }
    },
    "/api/v1/sla/analytics/performance/export": {
      "get": {
        "description": "Export performance aggregation data in JSON or CSV format.",
        "operationId": "export_performance_aggregation_endpoint_api_v1_sla_analytics_performance_export_get",
        "parameters": [
          {
            "description": "Export format: json or csv",
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "description": "Export format: json or csv",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "description": "Alias for site_id",
            "in": "query",
            "name": "site",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Alias for site_id",
              "title": "Site"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Performance Aggregation Endpoint",
        "tags": [
          "sla"
        ]
      }
    },
...
    "/api/v1/sla/analytics/snapshot/rebuild": {
      "post": {
        "description": "Rebuild analytics snapshot from live data (BE-025).\n\nThis endpoint:\n- Aggregates current SLA data from scratch\n- Creates a new snapshot row (preserves history)\n- Is idempotent - safe to call multiple times\n- Requires admin privileges\n\nUse this for reconciliation after migrations or data drift.",
        "operationId": "rebuild_analytics_snapshot_api_v1_sla_analytics_snapshot_rebuild_post",
        "parameters": [
          {
            "in": "query",
            "name": "snapshot_key",
            "required": false,
            "schema": {
              "default": "global",
              "title": "Snapshot Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SLAAnalyticsSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebuild Analytics Snapshot",
        "tags": [
          "sla"
        ]
      }
    },
...
  }
}